+++ src/main/java/guesthouse/oauth2/service/dto/KaKaoUserInfoResponse.java
@@ -1,4 +1,4 @@
-package guesthouse.oauth2.service;
+package guesthouse.oauth2.service.dto;
 
 public record KaKaoUserInfoResponse(Long id) {
 }
